requirements.txt
flask
binance-connector